requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
jinja2
python-multipart
python-jose
passlib[bcrypt]
pydantic-settings
reportlab
